requirements.txt
streamlit
openai
requests
pillow
gtts
pygame
speechrecognition
pyaudio
openai-whisper
vosk
firebase-admin
reportlab
plotly
pandas
numpy<2
python-dotenv
streamlit-option-menu
streamlit-aggrid 